fix: resolve 12 issues across security, correctness, docs, and cleanup
HIGH severity fixes:
- #364: Add mutex synchronization to TokenRouter in HTTP handler threads
- #365: Add graceful NPU/FLM worker shutdown with READY handshake
- #366: Add bounds checking in GGUF/h1b loaders to prevent OOM DoS
- #367: Add HIP_CHECK(hipGetLastError()) after all kernel launches

MEDIUM fixes:
- #374: Replace python3 -c with quoted heredocs in shell scripts
- #375: Move method bodies from inlined headers to .cpp files
- #410: Infer plugin tier from type (not hardcoded T2_GPU)

SECURITY fixes:
- #411: Redact exposed Stripe credentials from ROADMAP.md and store

DOCS/CLEANUP fixes:
- #412: Fix jammy→noble ROCm repos, Ubuntu 26.04→24.04, CI table
- #413: Fix stale paths in SECURITY.md
- #414: Fix CHANGELOG ordering, add 2026.07.16 entry, [[maybe_unused]]

Issues #401-408 covered by existing PR #415.
