+++ kernels/hadamard_rotate_butterfly.hip
@@ -29,9 +29,7 @@
 #include <hip/hip_fp16.h>
 #include <cstdint>
 #include <cmath>
-#ifndef HIP_CHECK
-#define HIP_CHECK(e) do { hipError_t _s = (e); if (_s != hipSuccess) { fprintf(stderr, "HIP Error %s:%d: %s\n", __FILE__, __LINE__, hipGetErrorString(_s)); abort(); } } while(0)
-#endif
+#include "hip_check.h"
 
 namespace {
