+++ kernels/rotorquant_pack.hip
@@ -23,9 +23,7 @@
 #include <cmath>
 
 #include "rocm_cpp/kv_rotorquant.h"
-#ifndef HIP_CHECK
-#define HIP_CHECK(e) do { hipError_t _s = (e); if (_s != hipSuccess) { fprintf(stderr, "HIP Error %s:%d: %s\n", __FILE__, __LINE__, hipGetErrorString(_s)); abort(); } } while(0)
-#endif
+#include "hip_check.h"
 
 namespace {
