+++ kernels/ternary_gemv.hip
@@ -26,9 +26,7 @@
 
 // LDS tile for activation vector — 512 floats shared across all warps
 #define LDS_TILE 512
-#ifndef HIP_CHECK
-#define HIP_CHECK(e) do { hipError_t _s = (e); if (_s != hipSuccess) { fprintf(stderr, "HIP Error %s:%d: %s\n", __FILE__, __LINE__, hipGetErrorString(_s)); abort(); } } while(0)
-#endif
+#include "hip_check.h"
 
 __attribute__((amdgpu_flat_work_group_size(BLOCK_SIZE, BLOCK_SIZE)))
 __global__ void ternary_gemv_kernel(
@@ -58,7 +56,8 @@ __global__ void ternary_gemv_kernel(
     for (int k_base = 0; k_base < K; k_base += LDS_TILE) {
         // Cooperative load of activation tile into LDS
         // 128 threads load 512 floats = 4 per thread
-        __syncthreads();
+        // Barrier needed for iterations > 0 to fence prior tile's warp_results
+        if (k_base > 0) __syncthreads();
         #pragma unroll
         for (int i = 0; i < LDS_TILE / BLOCK_SIZE; i++) {
             int k_idx = k_base + tid + i * BLOCK_SIZE;
@@ -156,7 +155,7 @@ __global__ void ternary_gemv_batched_kernel(
     float acc = 0.0f;
 
     for (int k_base = 0; k_base < K; k_base += LDS_TILE) {
-        __syncthreads();
+        if (k_base > 0) __syncthreads();
         #pragma unroll
         for (int i = 0; i < LDS_TILE / BLOCK_SIZE; i++) {
             int k_idx = k_base + tid + i * BLOCK_SIZE;
