+++ kernels/ternary_gemv_block_scaled.hip
@@ -11,9 +11,7 @@
 #define WARP_SIZE   32
 #define NUM_WARPS   (BLOCK_SIZE / WARP_SIZE)
 #define LDS_TILE_I8   2048
-#ifndef HIP_CHECK
-#define HIP_CHECK(e) do { hipError_t _s = (e); if (_s != hipSuccess) { fprintf(stderr, "HIP Error %s:%d: %s\n", __FILE__, __LINE__, hipGetErrorString(_s)); abort(); } } while(0)
-#endif
+#include "hip_check.h"
 
 __attribute__((amdgpu_flat_work_group_size(BLOCK_SIZE, BLOCK_SIZE)))
 __global__ void ternary_gemv_block_scaled_kernel(
