+++ kernels/ternary_gemv_sherry.hip
@@ -1,5 +1,5 @@
 #include <cstdio>
-#define HIP_CHECK(e) do { hipError_t _s = (e); if (_s != hipSuccess) { fprintf(stderr, "HIP Error %s:%d: %s\n", __FILE__, __LINE__, hipGetErrorString(_s)); abort(); } } while(0)
+#include "hip_check.h"
 
 // SPDX-License-Identifier: LicenseRef-PolyForm-Noncommercial-1.0.0
 // Sherry — see LICENSE-SHERRY.md and SHERRY-FILES.txt at the repo root.
