+++ kernels/zaya_moe_ternary_gemv.hip
@@ -1,5 +1,5 @@
 #include <cstdio>
-#define HIP_CHECK(e) do { hipError_t _s = (e); if (_s != hipSuccess) { fprintf(stderr, "HIP Error %s:%d: %s\n", __FILE__, __LINE__, hipGetErrorString(_s)); abort(); } } while(0)
+#include "hip_check.h"
 
 // Zaya MoE Ternary GEMV — expert-sparse ternary compute for Zaya's hybrid
 // CCA + MoE architecture with EDA router.
