+++ site/store/index.html
@@ -194,7 +194,7 @@ let stripe = null;
 const API_BASE = '';
 
 document.addEventListener('DOMContentLoaded', () => {
-  stripe = Stripe('pk_live_51TUvAmJnXt3bWED8N2qz5VRr63btuJjUeB8awZ6bJVWJvk0dWIgKBuXk56QRvUx4OItBsvkpooVZWwFgVcoJAuHp00aKGHfX7U');
+  stripe = Stripe('pk_live_REPLACE_WITH_YOUR_STRIPE_PUBLISHABLE_KEY');
   render();
 });
