+++ tests/bench_cb_compare.cpp
@@ -3,8 +3,7 @@
 #include <vector>
 #include "rocm_cpp/bonsai.h"
 
-#define HIP_CHECK(e) do { hipError_t _s = (e); if (_s != hipSuccess) { fprintf(stderr, "HIP Error %s:%d: %s\n", __FILE__, __LINE__, hipGetErrorString(_s)); abort(); } } while(0)
-#define HIP_OK(e) do { hipError_t _s = (e); if(_s!=hipSuccess) { fprintf(stderr,"HIP err %d at %s:%d\n",(int)_s,__FILE__,__LINE__); exit(1); }} while(0)
+#include "hip_check.h"
 extern "C" void bonsai_tq2_1024_cb_launch(const uint8_t*,const uint16_t*,uint16_t*,int,int,void*);
 
 constexpr int HS=2048,IS=6144,NL=28,NH=16,NKV=8,HD=128;
